+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.4] - TBD
+
+### Added
+- New `.significant_digits()` rounding method
+
 ## [1.1.3] - 2026-09-06
 
 ### Added